requirements.txt
python-telegram-bot==20.7
troposphere==4.5.3
